move seek table parser tests
